# openapi.yaml
openapi: 3.0.3
info:
  title: SocraticGemma API
  description: |
    ## SocraticGemma - Philosophy for Children AI Facilitator

    SocraticGemma is an API that uses Google's Gemma models to facilitate
    Socratic dialogues with children. It implements the Philosophy for Children
    (P4C) methodology, guiding children through philosophical inquiry by
    asking questions rather than giving answers.

    ### Features

    - **Socratic Dialogue**: Facilitates philosophical inquiry without giving direct answers
    - **Age-Appropriate**: Adapts language and concepts for different age groups (6-8, 9-12, 13-16)
    - **Evaluation**: Scores responses on socratism, age-fit, builds-on, openness, and advancement
    - **Forbidden Behavior Detection**: Identifies overhelp, lecture, correct, leading, and close behaviors
    - **RAG Enhancement**: Optional retrieval-augmented generation for context
    - **Baseline Comparison**: Compare P4C approach vs baseline helpful assistant

    ### API Usage

    1. Create a session with POST /sessions
    2. Send dialogue turns with POST /sessions/{id}/turns
    3. Get evaluation summaries with GET /sessions/{id}/eval

    Built for the P4C Hackathon 2024.
  version: 0.1.0
  contact:
    name: P4C Hackathon Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://your-production-url.com
    description: Production server
...
paths:
...
  /sessions:
    get:
      tags:
        - sessions
      summary: List all sessions
      description: Returns all non-expired sessions.
      operationId: list_sessions
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - sessions
      summary: Create a new Socratic dialogue session
      description: Creates a new session with the specified age group and stimulus.
      operationId: create_session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
            example:
              age_group: "9-12"
              stimulus:
                type: question
                content: "¿Por qué está mal mentir aunque nadie se entere?"
                title: "Mentira y honestidad"
              model_size: fast
              rag_enabled: false
              thinking_mode: true
              language: es
      responses:
        '201':
          description: Session created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id:
                    type: string
                    description: Unique session identifier
              example:
                session_id: "550e8400-e29b-41d4-a716-446655440000"
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sessions/{session_id}:
    get:
      tags:
        - sessions
      summary: Get session by ID
      description: Retrieves a session with full turn history.
      operationId: get_session
      parameters:
        - name: session_id
          in: path
          required: true
          description: The unique session identifier
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sessions/{session_id}/turns:
    post:
      tags:
        - dialogue
      summary: Process a dialogue turn with SSE streaming
      description: |
        Streams the model response as Server-Sent Events (SSE).
        
        Event types:
        - event:thinking - Thinking trace content chunks
        - event:token - Response token chunks  
        - event:complete - Final turn with evaluation scores
        - event:error - Error messages
      operationId: process_turn
      parameters:
        - name: session_id
          in: path
          required: true
          description: The unique session identifier
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - child_input
              properties:
                child_input:
                  type: string
                  description: The child's input or response
                  example: "Creo que está mal porque..."
      responses:
        '200':
          description: SSE stream of dialogue events
          content:
            text/event-stream:
              schema:
                type: string
                description: SSE event stream
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sessions/{session_id}/eval:
    get:
      tags:
        - evaluation
      summary: Get session evaluation summary
      description: Returns aggregated evaluation scores and question type distribution.
      operationId: get_evaluation_summary
      parameters:
        - name: session_id
          in: path
          required: true
          description: The unique session identifier
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Evaluation summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvalSummary'
              example:
                session_id: "550e8400-e29b-41d4-a716-446655440000"
                turn_count: 5
                avg_scores:
                  socratism: 4.2
                  age_fit: 4.5
                  builds_on: 4.0
                  openness: 4.3
                  advancement: 4.1
                  overall: 4.22
                  weighted_overall: 4.25
                question_type_distribution:
                  conceptual: 2
                  assumption: 1
                  evidence: 1
                  perspective: 1
                forbidden_behaviors_by_turn: []
        '400':
          description: Session has no turns
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /prompts:
    get:
      tags:
        - prompts
      summary: List available prompts
      description: Returns all available prompt templates.
      operationId: list_prompts
      responses:
        '200':
          description: List of prompts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PromptInfo'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /prompts/{name}:
    get:
      tags:
        - prompts
      summary: Get prompt by name
      description: Returns a specific prompt template by name.
      operationId: get_prompt
      parameters:
        - name: name
          in: path
          required: true
          description: The prompt name (without .yaml extension)
          schema:
            type: string
            enum:
              - baseline
              - system_base
              - age_6_8
              - age_9_12
              - age_13_16
              - question_types
              - forbidden_rules
              - output_format
      responses:
        '200':
          description: Prompt details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromptInfo'
        '404':
          description: Prompt not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rag/search:
    post:
      tags:
        - rag
      summary: Search P4C moves
      description: Search for relevant P4C question moves based on query text.
      operationId: search_moves
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RAGSearchRequest'
            example:
              query: "questions about fairness and justice"
              age_range: "9-12"
              top_k: 3
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RAGSearchResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: RAG service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rag/index:
    post:
      tags:
        - rag
      summary: Build RAG index
      description: Build the RAG index from the P4C moves data file.
      operationId: build_index
      responses:
        '200':
          description: Index built successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RAGIndexResponse'
              example:
                success: true
                message: Index built successfully
                move_count: 18
        '500':
          description: Build failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RAGIndexResponse'
...
components:
  schemas:
...
    Stimulus:
      type: object
      description: Stimulus that initiates a Socratic dialogue
      required:
        - type
        - content
      properties:
        type:
          type: string
          enum:
            - text
            - question
            - scenario
            - story
          description: Type of stimulus
        content:
          type: string
          description: The stimulus content (text, image URL, or scenario description)
        title:
          type: string
          description: Optional title for the stimulus
    CreateSessionRequest:
      type: object
      required:
        - age_group
        - stimulus
      properties:
        age_group:
          type: string
          enum:
            - "6-8"
            - "9-12"
            - "13-16"
          description: Age group of the child
        stimulus:
          $ref: '#/components/schemas/Stimulus'
        model_size:
          type: string
          enum:
            - fast
            - accurate
          default: fast
          description: Model size - 'fast' for 2B model, 'accurate' for 27B model
        rag_enabled:
          type: boolean
          default: false
          description: Whether to enable RAG for context augmentation
        thinking_mode:
          type: boolean
          default: true
          description: Whether to enable thinking mode (provider ordering)
        language:
          type: string
          default: en
          description: Language code for the dialogue
    EvalScores:
      type: object
      description: Evaluation scores for a turn response
      properties:
        socratism:
          type: number
          minimum: 1
          maximum: 5
          description: How well the response follows Socratic method (1-5)
        age_fit:
          type: number
          minimum: 1
          maximum: 5
          description: How appropriate for the child's age (1-5)
        builds_on:
          type: number
          minimum: 1
          maximum: 5
          description: How well it builds on child's previous response (1-5)
        openness:
          type: number
          minimum: 1
          maximum: 5
          description: How open-ended the question is (1-5)
        advancement:
          type: number
          minimum: 1
          maximum: 5
          description: How much it advances philosophical inquiry (1-5)
        overall:
          type: number
          minimum: 1
          maximum: 5
          description: Overall quality score (1-5)
        weighted_overall:
          type: number
          description: Weighted overall score
    TurnResponse:
      type: object
      description: Response containing a completed turn
      properties:
        id:
          type: integer
          description: Turn ID within the session
        content:
          type: string
          description: The model's response content
        question_type:
          type: string
          enum:
            - conceptual
            - assumption
            - evidence
            - perspective
            - implication
            - metacognitive
            - opening
            - statement
          description: Type of question asked
        thinking_trace:
          type: string
          description: The model's thinking trace (if enabled)
        eval_scores:
          $ref: '#/components/schemas/EvalScores'
        forbidden_behaviors_detected:
          type: array
          items:
            type: string
          description: List of forbidden behaviors detected (overhelp, lecture, correct, leading, close)
        rag_moves_used:
          type: array
          items:
            type: string
          description: RAG moves used for context augmentation
        timestamp:
          type: number
          description: Unix timestamp when the turn was created
    SessionResponse:
      type: object
      description: Full session response with history
      properties:
        session_id:
          type: string
          format: uuid
          description: Unique session identifier
        age_group:
          type: string
          description: Age group of the child
        stimulus:
          $ref: '#/components/schemas/Stimulus'
        model_size:
          type: string
          enum:
            - fast
            - accurate
          description: Model size used
        language:
          type: string
          description: Language code
        turns:
          type: array
          items:
            $ref: '#/components/schemas/TurnResponse'
        phases:
          type: object
          description: Phase completion status
          properties:
            stimulus:
              type: boolean
            questions:
              type: boolean
            agenda:
              type: boolean
            inquiry:
              type: boolean
            synthesis:
              type: boolean
        created_at:
          type: number
          description: Unix timestamp when session was created
    EvalScoresSummary:
      type: object
      description: Aggregated evaluation scores
      properties:
        socratism:
          type: number
          description: Average socratism score
        age_fit:
          type: number
          description: Average age fit score
        builds_on:
          type: number
          description: Average builds on score
        openness:
          type: number
          description: Average openness score
        advancement:
          type: number
          description: Average advancement score
        overall:
          type: number
          description: Average overall score
        weighted_overall:
          type: number
          description: Average weighted overall score
    EvalSummary:
      type: object
      description: Session evaluation summary with aggregated scores
      properties:
        session_id:
          type: string
          format: uuid
          description: Session identifier
        turn_count:
          type: integer
          description: Number of turns in the session
        avg_scores:
          $ref: '#/components/schemas/EvalScoresSummary'
        question_type_distribution:
          type: object
          additionalProperties:
            type: integer
          description: Count of each question type used
        forbidden_behaviors_by_turn:
          type: array
          items:
            type: object
            properties:
              turn_id:
                type: integer
              behaviors:
                type: array
                items:
                  type: string
          description: Forbidden behaviors detected per turn
...
    PromptInfo:
      type: object
      description: Information about a prompt template
      properties:
        name:
          type: string
          description: Prompt name/identifier
        description:
          type: string
          description: Brief description of the prompt
        content:
          type: string
          description: Full prompt content (YAML format)
...
    ErrorResponse:
      type: object
      description: Error response
      properties:
        error:
          type: string
          description: Error message
        detail:
          type: string
          description: Additional error details
    RAGSearchRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: Search query text
        age_range:
          type: string
          enum:
            - "6-8"
            - "9-12"
            - "13-16"
            - all
          description: Optional age range filter
        top_k:
          type: integer
          minimum: 1
          maximum: 20
          default: 3
          description: Number of results to return
    RAGSearchResponse:
      type: object
      description: Response schema for RAG search results
      properties:
        moves:
          type: array
          items:
            type: object
          description: List of matching P4C moves
        query:
          type: string
          description: Original query
        age_range:
          type: string
          description: Age range filter used
        count:
          type: integer
          description: Number of results
    RAGIndexResponse:
      type: object
      description: Response schema for RAG index build
      properties:
        success:
          type: boolean
        message:
          type: string
        move_count:
          type: integer
          description: Number of moves indexed
...